Update tft_synergy_calculator.m
Line 321
while incomplete_traits > 1 % Change this value to 0 to make the code give a better value, but it will take longer

I changed this back to what it was before, since if the value is 0, it can take a very long time.  With 1, you get almost the same result, and it's very fast.  I added the comment so that users can change the value if they want, but warned about it taking longer.